LayoutTests/fast/profiler/apply.html

<html>
<head>
<script src="resources/profiler-test-JS-resources.js"></script>
<script>
if (window.layoutTestController)
    layoutTestController.dumpAsText();

console.profile("Using the apply() method");
function startTest()
{
    var myObject = new fakeObject(1, 2);
    endTest();
}

function fakeObject (x, y)
{
    this.x = x;
    fakeInteriorFunction.apply(this, arguments);
}

function fakeInteriorFunction(x, y)
{
    this.y = y;
}

</script>
</head>

<body onload="startTest()">
This page's JavaScript has a call to apply() in it.
<br>
<br>
To run this test manually, load it in the browser then load the WebInspector and look at
the profile.  In the profile there should be a call to fakeInteriorFunction() and
a call to apply().
<div id="output"></div>
</body>
</html>
